Added Fish Bucket dispense behavior
